+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - added: Notification status analytics param
 - added: @bugsnag/react-native-performance
+- changed: Remove extra spaces and normal capitalization of mnemonic seed input
 - fixed: Some text inputs not selectable for some Samsung devices
 
 ## 4.1.0